+++ funcs/func_frame_trace.c
@@ -214,7 +214,7 @@ static void print_frame(struct ast_frame *frame)
 	switch (frame->frametype) {
 	case AST_FRAME_DTMF_END:
 		ast_verbose("FrameType: DTMF END\n");
-		ast_verbose("Digit: 0x%02X '%c'\n", frame->subclass.integer,
+		ast_verbose("Digit: 0x%02X '%c'\n", (unsigned)frame->subclass.integer,
 			frame->subclass.integer < ' ' ? ' ' : frame->subclass.integer);
 		break;
 	case AST_FRAME_VOICE:
@@ -390,7 +390,7 @@ static void print_frame(struct ast_frame *frame)
 		break;
 	case AST_FRAME_DTMF_BEGIN:
 		ast_verbose("FrameType: DTMF BEGIN\n");
-		ast_verbose("Digit: 0x%02X '%c'\n", frame->subclass.integer,
+		ast_verbose("Digit: 0x%02X '%c'\n", (unsigned)frame->subclass.integer,
 			frame->subclass.integer < ' ' ? ' ' : frame->subclass.integer);
 		break;
 	case AST_FRAME_BRIDGE_ACTION:
